First draft of a principals implementation and RedChannelsCollection.
User principals will go to principals/channels and group principals will
go to principals/collections. You can take a look at the principals when
you change the server line with the nodes array.
We need to find a way still to make this available. Also under /cloud?

Add a new collection class RedChannelsCollection which will be added to
the node tree list of the Sabre server object and return all available
channels.

Also add a baseUri to Sabre server.